UT-3577: Move File class
